refactor: create TimetableModal component
기존 페이지에 있는 Modal과 관련된 내용을 TimeTableModal으로 옮겼습니다.
