+++ src/pages/TimetablePage.tsx
@@ -3,13 +3,11 @@ import styled from 'styled-components';
 import {
   BlankLine,
   Heading1,
-  Heading2,
   Heading3,
   Input,
   Label,
   MainSideBarContainer,
   MediumButton,
-  Modal,
   SCREEN_SIZE,
   Select,
   showToast
@@ -20,6 +18,7 @@ import MainSideBar from '../components/MainSideBar';
 import { SubjectType } from '../types/Payload';
 import Api from '../api';
 import TimetableTable from '../components/Timetable/TimetableTable';
+import TimetableModal from '../components/Timetable/TimetableModal';
 
 const StyledContent = styled.div`
   margin: 3rem;
@@ -75,22 +74,12 @@ const TimetablePage: React.FC = () => {
 
   const open = useState<boolean>(false);
   const [isEditModal, setEditModal] = useState<boolean>(false);
-  const [modalInput, setModalInput] = useState<ModalInputState>({
+  const modalInput = useState<ModalInputState>({
     subjectName: '',
     teacher: '',
     zoom: ''
   });
   const [editId, setEditId] = useState<number | undefined>(undefined);
-
-  const onModalInputChange = (type: keyof ModalInputState) => (
-    e: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    e.persist();
-    setModalInput((current) => ({
-      ...current,
-      [type]: e.target.value
-    }));
-  };
 
   const fetchTimetable = useCallback(
     (page: number, _search: string) => {
@@ -113,7 +102,7 @@ const TimetablePage: React.FC = () => {
   useEffect(() => fetchTimetable(1, ''), [fetchTimetable]);
 
   const onCreateClick = () => {
-    setModalInput({
+    modalInput[1]({
       subjectName: '',
       teacher: '',
       zoom: ''
@@ -121,35 +110,9 @@ const TimetablePage: React.FC = () => {
     setEditModal(false);
     open[1](true);
   };
-  const onCreateModalButtonClick = async () => {
-    if (!modalInput.subjectName.trim()) {
-      showToast('과목명이 비어있습니다.', 'danger');
-      return;
-    }
-
-    if (!modalInput.teacher.trim()) {
-      showToast('담당 선생님이 비어있습니다.', 'danger');
-      return;
-    }
-
-    if (!modalInput.zoom.trim()) {
-      showToast('줌 링크가 비어있습니다.', 'danger');
-      return;
-    }
-
-    await Api.post('/timetable', {
-      subject: modalInput.subjectName.trim(),
-      teacher: modalInput.teacher.trim(),
-      url: modalInput.zoom
-    });
-
-    showToast('시간표를 추가했습니다.', 'success');
-    open[1](false);
-    fetchTimetable(1, search);
-  };
 
   const onEditClick = (id: number, currentData: SubjectType) => {
-    setModalInput({
+    modalInput[1]({
       subjectName: currentData.subject,
       teacher: currentData.teacher,
       zoom: currentData.url
@@ -157,20 +120,6 @@ const TimetablePage: React.FC = () => {
     setEditId(id);
     setEditModal(true);
     open[1](true);
-  };
-
-  const onEditModalButtonClick = async () => {
-    if (!editId) return;
-
-    await Api.put(`/timetable/${editId}`, {
-      subject: modalInput.subjectName.trim(),
-      teacher: modalInput.teacher.trim(),
-      url: modalInput.zoom
-    });
-
-    showToast('시간표를 수정했습니다.', 'success');
-    open[1](false);
-    fetchTimetable(1, search);
   };
 
   const onDeleteClick = async (id: number) => {
@@ -238,48 +187,13 @@ const TimetablePage: React.FC = () => {
         </StyledContent>
       </MainSideBarContainer>
 
-      <Modal width={450} height={450} name="AddTimetable" state={open}>
-        <Heading2>시간표 {isEditModal ? '수정하기' : '추가하기'}</Heading2>
-        <BlankLine gap={10} />
-
-        <Label>
-          과목명
-          <br />
-          (숫자, 띄어쓰기, 특수문자는 제외해주세요)
-        </Label>
-        <Input
-          type="text"
-          placeholder="ex. 수학"
-          value={modalInput.subjectName}
-          onChange={onModalInputChange('subjectName')}
-        />
-
-        <BlankLine gap={10} />
-
-        <Label>담당 선생님</Label>
-        <Input
-          type="text"
-          placeholder="ex. 홍길동"
-          value={modalInput.teacher}
-          onChange={onModalInputChange('teacher')}
-        />
-
-        <BlankLine gap={10} />
-
-        <Label>줌 링크</Label>
-        <Input
-          type="text"
-          placeholder="ex. https://zoom.us/j/1234?pwd=asdf"
-          value={modalInput.zoom}
-          onChange={onModalInputChange('zoom')}
-        />
-
-        <BlankLine gap={30} />
-
-        <MediumButton onClick={isEditModal ? onEditModalButtonClick : onCreateModalButtonClick}>
-          {isEditModal ? '수정하기' : '추가하기'}
-        </MediumButton>
-      </Modal>
+      <TimetableModal
+        open={open}
+        isEditModal={isEditModal}
+        editId={editId}
+        modalInput={modalInput}
+        fetchTimetable={fetchTimetable}
+      />
     </>
   );
 };
